Move prerender methods and replace unused hook
